+++ src/components/Root/index.js
@@ -1,1 +1,4 @@
-export { default } from './Root';
+import { hot } from 'react-hot-loader';
+import Root from './Root';
+
+export default hot(module)(Root);