+++ test/webpack/test-setup.spec.js
@@ -1,4 +1,4 @@
-// Test setup for Mocha tests
+// Test setup for webpack tests
 // Root-level Mocha hooks defined here apply to all tests regardless of file.
 
 const sinon = require('sinon');